+++ packages/engine/src/xr/XRSystem.ts
@@ -4,10 +4,10 @@ import XR8 from './8thwall/XR8'
 import { World } from './../ecs/classes/World'
 import XRAnchorSystem from './XRAnchorSystem'
 import XRCameraSystem from './XRCameraSystem'
-import XRControllerSystem from './XRControllerSystem'
 import XRDepthOcclusion from './XRDepthOcclusion'
 import XRDetectedPlanesSystem from './XRDetectedPlanesSystem'
 import XRHapticsSystem from './XRHapticsSystem'
+import XRInputSourceSystem from './XRInputSourceSystem'
 import XRLightProbeSystem from './XRLightProbeSystem'
 import XRPersistentAnchorSystem from './XRPersistentAnchorSystem'
 import XRScenePlacementShader from './XRScenePlacementShader'
@@ -60,7 +60,7 @@ export default async function XRSystem(world: World) {
       () => Promise.resolve({ default: XRPersistentAnchorSystem }),
       () => Promise.resolve({ default: XRAnchorSystem }),
       () => Promise.resolve({ default: XRCameraSystem }),
-      () => Promise.resolve({ default: XRControllerSystem }),
+      () => Promise.resolve({ default: XRInputSourceSystem }),
       () => Promise.resolve({ default: XRHapticsSystem }),
       () => Promise.resolve({ default: XRLightProbeSystem }),
       () => Promise.resolve({ default: XRDetectedPlanesSystem }),
